Correção de erros UserRegister

+++ src/main/java/com/example/Futurize/user/UserReponseDTO.java
@@ -1,6 +1,6 @@
 package com.example.Futurize.user;
 
-public record UserReponseDTO(Long id, String nome, String email, String senha) {
+public record UserReponseDTO(Integer id, String nome, String email, String senha) {
     public UserReponseDTO(User user){
         this(user.getId(), user.getNome(), user.getEmail(), user.getSenha());
     }
